requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-image>=0.18.0
opencv-python>=4.5.0
tqdm>=4.60.0
pillow>=8.2.0
scipy>=1.7.0
pandas>=1.3.0
timm>=0.4.12
monai>=0.9.0
tensorboard>=2.5.0
pycocotools>=2.0.2
safetensors>=0.3.1
transformers>=4.12.0 